update item-submission.xml
Update the submission processes and collection mapping to accommodate
the separation of the rights-metadata entry into a discrete step.

All collections with a non-default input form have been mapped to a
submission-process WITHOUT the rights-metadata entry step because it
is assumed that rights entry will be part of the non-default input
form.  This was done so as not to change the UX for existing users.